requirements.txt
streamlit
langgraph
langchain
langchain-community
langchain-google-genai
google-generativeai
python-dotenv
requests
